refactor(Editor, ECS, Scene): Improve component management and simplify modal pop-up interfaces
- Updated VerticalLayout.cpp to enhance precision in spacing calculations by using float division.
- Simplified BaseModalPopUp's open function in BaseWindow.cpp and removed button callback functionality in BaseWindow.h, reducing complexity.
- Restructured ModalAssetsSearchPopUp and ModalECSSearchPopUp classes to improve clarity and streamline their implementations by consolidating methods and removing redundancy.
- Added removeAllChildren method in BaseComponent to enhance child component management and modified Scene.cpp to ensure component names are set based on provided values, improving overall functionality and reliability.

+++ sources/Editor/GuiComponents/VerticalLayout.cpp
@@ -221,7 +221,7 @@ namespace NX::Gui
                 }
                 space -= child->unsafeCastTo<Widget>()->getHeight();
             }
-            space /= _children.size() - 1ll;
+            space /= static_cast<float>(_children.size() - 1ll);
         }
 
         if (_paddings.z != 0.f)
